requirements.txt
fire
sentencepiece
simple-parsing
bitsandbytes
scipy
datasets
wandb
tensorboard
# triton==2.2.0  # also works without but throws warning and is slower
llama-index
streamlit
flask==2.1.3
llama-index-llms-mistralai
llama-index-vector-stores-weaviate
ipython
llama-index-embeddings-mistralai